+++ routes/api.py
@@ -1,3 +1,6 @@
+# /routes/api.py
+# by Arnaud Cresp - 2025
+
 from flask import Blueprint, request, jsonify
 from services.data_loader import load_nodes
 from services.nmos_connection import change_source, disconnect_receiver