+++ services/data_loader.py
@@ -1,3 +1,6 @@
+# /services/data_loader.py
+# by Arnaud Cresp - 2025
+
 import os
 import json
 import requests